storage: added VariantStatsManager to modulate stats. not tested.
It would allow creating the stats in thread batches, or change the
VariantStatsCalculator for a JNI one, etc.
